requirements.txt
pytest==7.4.0
pytest-playwright==0.4.3
playwright==1.30.0
requests==2.31.0
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.0